cli/internal/oauth: move to top-level "internal"
