+++ internal/oauth/manager/manager.go
@@ -11,9 +11,9 @@ import (
 
 	"github.com/docker/cli/cli/config/credentials"
 	"github.com/docker/cli/cli/config/types"
-	"github.com/docker/cli/cli/internal/oauth"
-	"github.com/docker/cli/cli/internal/oauth/api"
 	"github.com/docker/cli/cli/streams"
+	"github.com/docker/cli/internal/oauth"
+	"github.com/docker/cli/internal/oauth/api"
 	"github.com/docker/cli/internal/tui"
 	"github.com/docker/docker/registry"
 	"github.com/morikuni/aec"
