+++ internal/oauth/manager/util.go
